+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "concordium-grpc-api"]
 	path = concordium-grpc-api
-	url = https://github.com/Concordium/concordium-grpc-api.git
+	url = ../concordium-grpc-api.git
