--- Sources/Concordium/Domain/GRPCError.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-public enum GRPCError: Error, Equatable {
-    case unexpectedBase58CheckVersion(expected: UInt8, actual: UInt8)
-    case missingRequiredValue(String)
-    case unsupportedValue(String)
-    case valueOutOfBounds
-}
