+++ Sources/Concordium/Domain/Queries.swift
@@ -1,5 +1,18 @@
+import ConcordiumWalletCrypto
 import Foundation
 import NIO
+
+public typealias CryptographicParameters = ConcordiumWalletCrypto.GlobalContext
+
+extension CryptographicParameters: FromGRPC {
+    static func fromGRPC(_ grpc: Concordium_V2_CryptographicParameters) -> Self {
+        .init(
+            onChainCommitmentKey: grpc.onChainCommitmentKey,
+            bulletproofGenerators: grpc.bulletproofGenerators,
+            genesisString: grpc.genesisString
+        )
+    }
+}
 
 /// Describes the possible status variants of a transaction submitted to a node.
 public enum TransactionStatus {
