+++ Sources/Concordium/Integration/WalletProxy.swift
