+++ Sources/Concordium/Wallet/WalletIdentity.swift
